small bug fix in light of test harness fail - th
git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@1434 25464f21-6099-402c-b30a-640f1d40d38b

+++ torus/vector_mod.f90
@@ -55,9 +55,9 @@ module vector_mod
   ! The definition of the vector type
 
   type VECTOR
-     real(single) :: x
-     real(single) :: y
-     real(single) :: z
+     real(double) :: x
+     real(double) :: y
+     real(double) :: z
   end type VECTOR
 
   type doubleVector
